+++ clients/include/blas3/testing_herkx_strided_batched.hpp
@@ -431,9 +431,9 @@ void testing_herkx_strided_batched(const Arguments& arg)
     hipblasFillMode_t  uplo     = char2hipblas_fill(arg.uplo);
     hipblasOperation_t transA   = char2hipblas_operation(arg.transA);
     int64_t            K1       = (transA == HIPBLAS_OP_N ? K : N);
-    hipblasStride      stride_A = size_t(lda) * K1 * stride_scale;
-    hipblasStride      stride_B = size_t(ldb) * K1 * stride_scale;
-    hipblasStride      stride_C = size_t(ldc) * N * stride_scale;
+    hipblasStride      stride_A = lda * K1 * stride_scale;
+    hipblasStride      stride_B = ldb * K1 * stride_scale;
+    hipblasStride      stride_C = ldc * N * stride_scale;
     size_t             A_size   = stride_A * batch_count;
     size_t             B_size   = stride_B * batch_count;
     size_t             C_size   = stride_C * batch_count;
