+++ clients/include/blas3/testing_syrk_strided_batched.hpp
@@ -341,8 +341,8 @@ void testing_syrk_strided_batched(const Arguments& arg)
     int64_t            batch_count  = arg.batch_count;
 
     int64_t       K1       = (transA == HIPBLAS_OP_N ? K : N);
-    hipblasStride stride_A = size_t(lda) * K1 * stride_scale;
-    hipblasStride stride_C = size_t(ldc) * N * stride_scale;
+    hipblasStride stride_A = lda * K1 * stride_scale;
+    hipblasStride stride_C = ldc * N * stride_scale;
     size_t        A_size   = stride_A * batch_count;
     size_t        C_size   = stride_C * batch_count;
