+++ clients/include/blas3/testing_syrkx_strided_batched.hpp
@@ -407,9 +407,9 @@ void testing_syrkx_strided_batched(const Arguments& arg)
     T h_beta  = arg.get_beta<T>();
 
     int64_t       K1       = transA == HIPBLAS_OP_N ? K : N;
-    hipblasStride stride_A = size_t(lda) * K1 * stride_scale;
-    hipblasStride stride_B = size_t(ldb) * K1 * stride_scale;
-    hipblasStride stride_C = size_t(ldc) * N * stride_scale;
+    hipblasStride stride_A = lda * K1 * stride_scale;
+    hipblasStride stride_B = ldb * K1 * stride_scale;
+    hipblasStride stride_C = ldc * N * stride_scale;
 
     hipblasLocalHandle handle(arg);
